requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
libsql-client==0.3.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
python-dotenv==1.0.0
redis==5.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
yarl==1.9.4
